Fixes issue on move when using Touch. Moved item is no longer lost.
Beacuse touch uses a timer called from the starting element's draggr
object, the "this" in the drop is always going to be the starting
element's event. This mean the starting (vue-draggr.onAdd) was being
called for the originating vue component. This needs to come from the
prospective parent. Simple change to the call, so it always originates
from the parentEL fixes the issue on touch. Desktop is unaffected.